+++ apps/decodex/src/agent/app_server/tests/schema_tests.rs
@@ -1,6 +1,5 @@
 use std::fs;
 
-use serde_json;
 use tempfile::TempDir;
 
 use crate::agent::app_server::{
